share host-session wire shapes through agent-bundle/contracts; drop the Workbench shim
